% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelets.R
\name{MexicanHat}
\alias{MexicanHat}
\title{Mexican Hat}
\description{
Mexican Hat

Mexican Hat
}
\details{
This is a derivative of a Gaussian of order 2.
}
\section{Super class}{
\code{torchwavelets::Derivative of Gaussian} -> \code{Mexican Hat}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Mexican Hat-new}{\code{MexicanHat$new()}}
\item \href{#method-Mexican Hat-clone}{\code{MexicanHat$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="torchwavelets" data-topic="Derivative of Gaussian" data-id="coi"><a href='../../torchwavelets/html/Derivative of Gaussian.html#method-Derivative of Gaussian-coi'><code>torchwavelets::Derivative of Gaussian$coi()</code></a></span></li>
<li><span class="pkg-link" data-pkg="torchwavelets" data-topic="Derivative of Gaussian" data-id="fourier_period"><a href='../../torchwavelets/html/Derivative of Gaussian.html#method-Derivative of Gaussian-fourier_period'><code>torchwavelets::Derivative of Gaussian$fourier_period()</code></a></span></li>
<li><span class="pkg-link" data-pkg="torchwavelets" data-topic="Derivative of Gaussian" data-id="frequency"><a href='../../torchwavelets/html/Derivative of Gaussian.html#method-Derivative of Gaussian-frequency'><code>torchwavelets::Derivative of Gaussian$frequency()</code></a></span></li>
<li><span class="pkg-link" data-pkg="torchwavelets" data-topic="Derivative of Gaussian" data-id="scale_from_period"><a href='../../torchwavelets/html/Derivative of Gaussian.html#method-Derivative of Gaussian-scale_from_period'><code>torchwavelets::Derivative of Gaussian$scale_from_period()</code></a></span></li>
<li><span class="pkg-link" data-pkg="torchwavelets" data-topic="Derivative of Gaussian" data-id="time"><a href='../../torchwavelets/html/Derivative of Gaussian.html#method-Derivative of Gaussian-time'><code>torchwavelets::Derivative of Gaussian$time()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mexican Hat-new"></a>}}
\if{latex}{\out{\hypertarget{method-Mexican Hat-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MexicanHat$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mexican Hat-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Mexican Hat-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MexicanHat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
