% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelets.R
\name{Morlet}
\alias{Morlet}
\title{Complex Morlet wavelet, centred at zero.}
\description{
Complex Morlet wavelet, centred at zero.

Complex Morlet wavelet, centred at zero.
}
\details{
The standard version is:
\deqn{pi**-0.25 * exp(1j*w*x) * exp(-0.5*(x**2))}
This commonly used wavelet is often referred to simply as the Morlet wavelet.
Note that this simplified version can cause admissibility problems at low values
of `w`.
The complete version is:
\deqn{pi**-0.25 * (exp(1j*w*x) - exp(-0.5*(w**2))) * exp(-0.5*(x**2))}
The complete version of the Morlet wavelet, with a correction term to improve
admissibility. For `w` greater than 5, the correction term is negligible
Note that the energy of the return wavelet is not normalised according to `s`.
The fundamental frequency of this wavelet in Hz is given by `f = 2*s*w*r / M`
where r is the sampling rate.
}
\seealso{
scipy.signal.gausspulse
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Morlet-new}{\code{Morlet$new()}}
\item \href{#method-Morlet-time}{\code{Morlet$time()}}
\item \href{#method-Morlet-fourier_period}{\code{Morlet$fourier_period()}}
\item \href{#method-Morlet-scale_from_period}{\code{Morlet$scale_from_period()}}
\item \href{#method-Morlet-frequency}{\code{Morlet$frequency()}}
\item \href{#method-Morlet-coi}{\code{Morlet$coi()}}
\item \href{#method-Morlet-clone}{\code{Morlet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morlet-new"></a>}}
\if{latex}{\out{\hypertarget{method-Morlet-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morlet$new(w0 = 6)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{w0}}{the nondimensional frequency constant. If this is set too low
then the wavelet does not sample very well: a value over 5 should be ok;
Terrence and Compo set it to 6.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morlet-time"></a>}}
\if{latex}{\out{\hypertarget{method-Morlet-time}{}}}
\subsection{Method \code{time()}}{
Value of the wavelet at the given time
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morlet$time(t, s = 1, complete = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{Time. If s is not specified, this can be used as the non-dimensional time t/s.}

\item{\code{s}}{Scaling factor. Default is 1.}

\item{\code{complete}}{Whether to use the compxlete or the standard version.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morlet-fourier_period"></a>}}
\if{latex}{\out{\hypertarget{method-Morlet-fourier_period}{}}}
\subsection{Method \code{fourier_period()}}{
Equivalent Fourier period of Morlet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morlet$fourier_period(s)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{Scaling factor}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morlet-scale_from_period"></a>}}
\if{latex}{\out{\hypertarget{method-Morlet-scale_from_period}{}}}
\subsection{Method \code{scale_from_period()}}{
Compute the scale from the fourier period
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morlet$scale_from_period(p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{Fourier period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morlet-frequency"></a>}}
\if{latex}{\out{\hypertarget{method-Morlet-frequency}{}}}
\subsection{Method \code{frequency()}}{
Frequency representation of Morlet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morlet$frequency(w, s = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{w}}{angular frequency. If s is not specified, i.e. set to 1,
this can be used as the non-dimensional angular frequency w * s.}

\item{\code{s}}{the scaling factor. Default is 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morlet-coi"></a>}}
\if{latex}{\out{\hypertarget{method-Morlet-coi}{}}}
\subsection{Method \code{coi()}}{
The e folding time for the autocorrelation of wavelet
power at each scale, i.e. the timescale over which an edge
effect decays by a factor of 1/e^2.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morlet$coi(s)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{Scaling factor}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morlet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Morlet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morlet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
