% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morlet.R
\name{morlet_fourier}
\alias{morlet_fourier}
\title{The Morlet wavelet, in Fourier domain}
\usage{
morlet_fourier(s, omega_a, omega)
}
\arguments{
\item{s}{scale parameter for this wavelet (a scalar).}

\item{omega_a}{the analysis frequency to be used (a scalar). Must be the DFT bin, not a frequency in Hertz.}

\item{omega}{a vector of frequencies to be considered.  Must be the DFT bin, not a frequency in Hertz.}
}
\description{
The Morlet wavelet, in Fourier domain
}
