% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelets.R
\name{Paul}
\alias{Paul}
\title{Paul wavelet}
\description{
Paul wavelet

Paul wavelet
}
\details{
The Paul wavelet is defined as
\deqn{(2 ** m * i ** m * m!) / (pi * (2 * m)!) * (1 - i * t / s) ** -(m + 1)}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Paul-new}{\code{Paul$new()}}
\item \href{#method-Paul-time}{\code{Paul$time()}}
\item \href{#method-Paul-fourier_period}{\code{Paul$fourier_period()}}
\item \href{#method-Paul-scale_from_period}{\code{Paul$scale_from_period()}}
\item \href{#method-Paul-frequency}{\code{Paul$frequency()}}
\item \href{#method-Paul-coi}{\code{Paul$coi()}}
\item \href{#method-Paul-clone}{\code{Paul$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Paul-new"></a>}}
\if{latex}{\out{\hypertarget{method-Paul-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Paul$new(m = 4)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{m}}{the order}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Paul-time"></a>}}
\if{latex}{\out{\hypertarget{method-Paul-time}{}}}
\subsection{Method \code{time()}}{
Value of the wavelet at the given time
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Paul$time(t, s = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{Time. If s is not specified, this can be used as the non-dimensional time t/s.}

\item{\code{s}}{Scaling factor. Default is 1.}

\item{\code{complete}}{Whether to use the complete or the standard version.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Paul-fourier_period"></a>}}
\if{latex}{\out{\hypertarget{method-Paul-fourier_period}{}}}
\subsection{Method \code{fourier_period()}}{
Equivalent Fourier period of wavelet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Paul$fourier_period(s)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{Scaling factor}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Paul-scale_from_period"></a>}}
\if{latex}{\out{\hypertarget{method-Paul-scale_from_period}{}}}
\subsection{Method \code{scale_from_period()}}{
Compute the scale from the fourier period
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Paul$scale_from_period(period)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{period}}{Fourier period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Paul-frequency"></a>}}
\if{latex}{\out{\hypertarget{method-Paul-frequency}{}}}
\subsection{Method \code{frequency()}}{
Frequency representation of wavelet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Paul$frequency(w, s = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{w}}{angular frequency. If s is not specified, i.e. set to 1,
this can be used as the non-dimensional angular frequency w * s.}

\item{\code{s}}{the scaling factor. Default is 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Paul-coi"></a>}}
\if{latex}{\out{\hypertarget{method-Paul-coi}{}}}
\subsection{Method \code{coi()}}{
The e folding time for the autocorrelation of wavelet
power at each scale, i.e. the timescale over which an edge
effect decays by a factor of 1/e^2.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Paul$coi(s)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{Scaling factor}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Paul-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Paul-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Paul$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
