% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{plot_wavelet_diagram}
\alias{plot_wavelet_diagram}
\title{plot a wavelet diagram (scaleogram)}
\usage{
plot_wavelet_diagram(x, freqs, grid, s, fs, f_end, type = "magnitude")
}
\arguments{
\item{x}{the signal}

\item{freqs}{the analysis frequencies used}

\item{grid}{the wavelet transform, a time-frequency grid}

\item{s}{the wavelet scale used}

\item{fs}{the sampling rate}

\item{f_end}{the highest analysis frequency to use}

\item{type}{what to plot; one of "Magnitude", "Magnitude squared", or "Magnitude (square root)"}
}
\description{
plot a wavelet diagram (scaleogram)
}
