% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{wavelet_grid}
\alias{wavelet_grid}
\title{perform wavelet transform on a grid of analysis frequencies}
\usage{
wavelet_grid(x, s, f_start, f_end, fs)
}
\arguments{
\item{x}{the signal}

\item{s}{the wavelet scale}

\item{f_start}{the lowest analysis frequency to use}

\item{f_end}{the highest analysis frequency to use}

\item{fs}{the sampling rate}
}
\value{
a list of transform results and analysis frequencies used
}
\description{
perform wavelet transform on a grid of analysis frequencies
}
