% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{wavelet_transform}
\alias{wavelet_transform}
\title{Wrapper constructor for the continuous wavelet transform as described in Torrence and Compo,
"A Practical Guide to Wavelet Analysis".}
\usage{
wavelet_transform(
  signal_length,
  dt = 1,
  dj = 0.125,
  wavelet = torchwavelets::Morlet$new(),
  fourier = TRUE
)
}
\arguments{
\item{signal_length}{length of the signal to be processed}

\item{dt}{sample spacing, default is 1}

\item{dj}{scale distribution parameter, default is 0.125}

\item{wavelet}{wavelet object}

\item{fourier}{whether to run in the frequency domain}
}
\description{
Wrapper constructor for the continuous wavelet transform as described in Torrence and Compo,
"A Practical Guide to Wavelet Analysis".
}
\seealso{
WaveletTransform
}
