% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{wavelet_transform_fourier}
\alias{wavelet_transform_fourier}
\title{The wavelet transform, computed in the Fourier domain
Expects that the wavelet has already been constructed.}
\usage{
wavelet_transform_fourier(x, m)
}
\arguments{
\item{x}{the signal, a vector (in the time domain)}

\item{m}{the wavelet to use (in Fourier representation)}
}
\value{
the wavelet transform, in the time domain
}
\description{
The wavelet transform, computed in the Fourier domain
Expects that the wavelet has already been constructed.
}
