% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{wavelet_transform_from_specs}
\alias{wavelet_transform_from_specs}
\title{The wavelet transform, computed in the Fourier domain
Constructs the wavelet for a desired scale and analysis frequency.}
\usage{
wavelet_transform_from_specs(type, x, omega_a, s, fs)
}
\arguments{
\item{type}{the type of wavelet. Currently has to be "morlet".}

\item{x}{the signal, a vector (in the time domain)}

\item{omega_a}{the analysis frequency to be used, in Hertz.}

\item{s}{the scale parameter to be used with the wavelets to be constructed.}

\item{fs}{the sampling frequency in Hertz.}
}
\value{
the wavelet transform, in the time domain
}
\description{
The wavelet transform, computed in the Fourier domain
Constructs the wavelet for a desired scale and analysis frequency.
}
