% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{WaveletTransform}
\alias{WaveletTransform}
\title{Continuous wavelet transform as described in Torrence and Compo,
"A Practical Guide to Wavelet Analysis".}
\description{
Continuous wavelet transform as described in Torrence and Compo,
"A Practical Guide to Wavelet Analysis".

Continuous wavelet transform as described in Torrence and Compo,
"A Practical Guide to Wavelet Analysis".
}
\details{
Like _QUVA-Lab/PyTorchWavelets_, this does not use the FFT,
but a filter bank consisting of `torch::nn_conv1d()` modules.
The code expects input sequences to be one-dimensional. It works with batches
of signals.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WaveletTransform-new}{\code{WaveletTransform$new()}}
\item \href{#method-WaveletTransform-build_filters}{\code{WaveletTransform$build_filters()}}
\item \href{#method-WaveletTransform-filter_bank}{\code{WaveletTransform$filter_bank()}}
\item \href{#method-WaveletTransform-cwt_time}{\code{WaveletTransform$cwt_time()}}
\item \href{#method-WaveletTransform-cwt_freq}{\code{WaveletTransform$cwt_freq()}}
\item \href{#method-WaveletTransform-cwt}{\code{WaveletTransform$cwt()}}
\item \href{#method-WaveletTransform-compute_optimal_scales}{\code{WaveletTransform$compute_optimal_scales()}}
\item \href{#method-WaveletTransform-power}{\code{WaveletTransform$power()}}
\item \href{#method-WaveletTransform-get_padding}{\code{WaveletTransform$get_padding()}}
\item \href{#method-WaveletTransform-compute_minimum_scale}{\code{WaveletTransform$compute_minimum_scale()}}
\item \href{#method-WaveletTransform-fourier_period}{\code{WaveletTransform$fourier_period()}}
\item \href{#method-WaveletTransform-scale_from_period}{\code{WaveletTransform$scale_from_period()}}
\item \href{#method-WaveletTransform-fourier_periods}{\code{WaveletTransform$fourier_periods()}}
\item \href{#method-WaveletTransform-fourier_frequencies}{\code{WaveletTransform$fourier_frequencies()}}
\item \href{#method-WaveletTransform-coi}{\code{WaveletTransform$coi()}}
\item \href{#method-WaveletTransform-complex_wavelet}{\code{WaveletTransform$complex_wavelet()}}
\item \href{#method-WaveletTransform-clone}{\code{WaveletTransform$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaveletTransform-new"></a>}}
\if{latex}{\out{\hypertarget{method-WaveletTransform-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaveletTransform$new(
  signal_length,
  dt = 1,
  dj = 0.125,
  wavelet = torchwavelets::Morlet$new(),
  fourier = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signal_length}}{length of the signal to be processed}

\item{\code{dt}}{sample spacing, default is 1}

\item{\code{dj}}{scale distribution parameter, default is 0.125}

\item{\code{wavelet}}{wavelet object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaveletTransform-build_filters"></a>}}
\if{latex}{\out{\hypertarget{method-WaveletTransform-build_filters}{}}}
\subsection{Method \code{build_filters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaveletTransform$build_filters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaveletTransform-filter_bank"></a>}}
\if{latex}{\out{\hypertarget{method-WaveletTransform-filter_bank}{}}}
\subsection{Method \code{filter_bank()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaveletTransform$filter_bank(filters, padding_type = "same")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaveletTransform-cwt_time"></a>}}
\if{latex}{\out{\hypertarget{method-WaveletTransform-cwt_time}{}}}
\subsection{Method \code{cwt_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaveletTransform$cwt_time(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaveletTransform-cwt_freq"></a>}}
\if{latex}{\out{\hypertarget{method-WaveletTransform-cwt_freq}{}}}
\subsection{Method \code{cwt_freq()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaveletTransform$cwt_freq(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaveletTransform-cwt"></a>}}
\if{latex}{\out{\hypertarget{method-WaveletTransform-cwt}{}}}
\subsection{Method \code{cwt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaveletTransform$cwt(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaveletTransform-compute_optimal_scales"></a>}}
\if{latex}{\out{\hypertarget{method-WaveletTransform-compute_optimal_scales}{}}}
\subsection{Method \code{compute_optimal_scales()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaveletTransform$compute_optimal_scales()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaveletTransform-power"></a>}}
\if{latex}{\out{\hypertarget{method-WaveletTransform-power}{}}}
\subsection{Method \code{power()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaveletTransform$power(x, unbias = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaveletTransform-get_padding"></a>}}
\if{latex}{\out{\hypertarget{method-WaveletTransform-get_padding}{}}}
\subsection{Method \code{get_padding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaveletTransform$get_padding(padding_type, kernel_size)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaveletTransform-compute_minimum_scale"></a>}}
\if{latex}{\out{\hypertarget{method-WaveletTransform-compute_minimum_scale}{}}}
\subsection{Method \code{compute_minimum_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaveletTransform$compute_minimum_scale()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaveletTransform-fourier_period"></a>}}
\if{latex}{\out{\hypertarget{method-WaveletTransform-fourier_period}{}}}
\subsection{Method \code{fourier_period()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaveletTransform$fourier_period(s)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaveletTransform-scale_from_period"></a>}}
\if{latex}{\out{\hypertarget{method-WaveletTransform-scale_from_period}{}}}
\subsection{Method \code{scale_from_period()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaveletTransform$scale_from_period(p)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaveletTransform-fourier_periods"></a>}}
\if{latex}{\out{\hypertarget{method-WaveletTransform-fourier_periods}{}}}
\subsection{Method \code{fourier_periods()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaveletTransform$fourier_periods()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaveletTransform-fourier_frequencies"></a>}}
\if{latex}{\out{\hypertarget{method-WaveletTransform-fourier_frequencies}{}}}
\subsection{Method \code{fourier_frequencies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaveletTransform$fourier_frequencies()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaveletTransform-coi"></a>}}
\if{latex}{\out{\hypertarget{method-WaveletTransform-coi}{}}}
\subsection{Method \code{coi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaveletTransform$coi(t_min, t_max)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaveletTransform-complex_wavelet"></a>}}
\if{latex}{\out{\hypertarget{method-WaveletTransform-complex_wavelet}{}}}
\subsection{Method \code{complex_wavelet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaveletTransform$complex_wavelet(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaveletTransform-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WaveletTransform-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaveletTransform$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
