% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morlet.R
\name{morlet}
\alias{morlet}
\title{The Morlet wavelet, in the time domain}
\usage{
morlet(omega, s, t_k, t)
}
\arguments{
\item{omega}{the analysis frequency to be used (a scalar, in Hertz).}

\item{s}{scale parameter for this wavelet (a scalar).}

\item{t_k}{location parameter for this wavelet (a scalar).}

\item{t}{a vector of measurement times.}
}
\value{
a vector of wavelet values (complex).
}
\description{
\deqn{\Psi_{\omega_{a},s,t_{k}}(t_n) = (e^{-i \omega_{a} (t_n - t_k)} - e^{-s^2}) \ e^{- \omega_a^2 (t_n - t_k )^2 /(2s )^2}}
}
\details{
See Vistnes, Physics of Oscillations and Waves, eq. (14.8).
}
