% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelets.R
\name{Morlet}
\alias{Morlet}
\title{Complex Morlet wavelet, centered at zero.}
\description{
Complex Morlet wavelet, centered at zero.

Complex Morlet wavelet, centered at zero.
}
\details{
The standard version is:
\deqn{pi**-0.25 * exp(1j*w*x) * exp(-0.5*(x**2))}
This commonly used wavelet is often referred to simply as the Morlet wavelet.
Note that this simplified version can cause admissibility problems at low values
of `w`.
The complete version is:
\deqn{pi**-0.25 * (exp(1j*w*x) - exp(-0.5*(w**2))) * exp(-0.5*(x**2))}
The complete version of the Morlet wavelet, with a correction term to improve
admissibility. For `w` greater than 5, the correction term is negligible.
Note that the energy of the returned wavelet is not normalized according to `s`.
The fundamental frequency of this wavelet in Hz is given by `f = 2*s*w*r / M`,
where r is the sampling rate.
}
\section{Super class}{
\code{\link[torchwavelets:Wavelet]{torchwavelets::Wavelet}} -> \code{Morlet}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{is_complex}}{whether the wavelet representation in the time domain is complex}

\item{\code{w0}}{the non-dimensional frequency constant.}

\item{\code{C_d}}{value of C_d from TC98}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Morlet-new}{\code{Morlet$new()}}
\item \href{#method-Morlet-time}{\code{Morlet$time()}}
\item \href{#method-Morlet-fourier_period}{\code{Morlet$fourier_period()}}
\item \href{#method-Morlet-scale_from_period}{\code{Morlet$scale_from_period()}}
\item \href{#method-Morlet-frequency}{\code{Morlet$frequency()}}
\item \href{#method-Morlet-coi}{\code{Morlet$coi()}}
\item \href{#method-Morlet-clone}{\code{Morlet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morlet-new"></a>}}
\if{latex}{\out{\hypertarget{method-Morlet-new}{}}}
\subsection{Method \code{new()}}{
save `w0` (and possibly set `C_d`)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morlet$new(w0 = 6)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{w0}}{the non-dimensional frequency constant. If this is set too low,
then the wavelet does not sample very well: a value higher than 5 should be ok.
Torrence and Compo set it to 6.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morlet-time"></a>}}
\if{latex}{\out{\hypertarget{method-Morlet-time}{}}}
\subsection{Method \code{time()}}{
value of the wavelet at the given times
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morlet$time(t, s = 1, complete = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{time. If `s` is not specified, this can be used as the non-dimensional
time t/s.}

\item{\code{s}}{scaling factor. Default is 1.}

\item{\code{complete}}{whether to use the complete or the standard version.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morlet-fourier_period"></a>}}
\if{latex}{\out{\hypertarget{method-Morlet-fourier_period}{}}}
\subsection{Method \code{fourier_period()}}{
equivalent Fourier period
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morlet$fourier_period(s)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{scaling factor}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morlet-scale_from_period"></a>}}
\if{latex}{\out{\hypertarget{method-Morlet-scale_from_period}{}}}
\subsection{Method \code{scale_from_period()}}{
compute the scale from the fourier period
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morlet$scale_from_period(p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{Fourier period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morlet-frequency"></a>}}
\if{latex}{\out{\hypertarget{method-Morlet-frequency}{}}}
\subsection{Method \code{frequency()}}{
frequency representation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morlet$frequency(w, s = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{w}}{angular frequency. If `s` is not specified, i.e. set to 1,
this can be used as the non-dimensional angular frequency w * s.}

\item{\code{s}}{the scaling factor. Default is 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morlet-coi"></a>}}
\if{latex}{\out{\hypertarget{method-Morlet-coi}{}}}
\subsection{Method \code{coi()}}{
The e-folding time for the autocorrelation of wavelet
power at each scale, i.e. the timescale over which an edge
effect decays by a factor of 1/e^2.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morlet$coi(s)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{scaling factor}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morlet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Morlet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morlet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
