% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{wavelet_transform}
\alias{wavelet_transform}
\title{Continuous wavelet transform as described in Torrence and Compo,
"A Practical Guide to Wavelet Analysis".}
\usage{
wavelet_transform(
  signal_length,
  dt = 1,
  dj = 0.125,
  wavelet = torchwavelets::Morlet$new(),
  unbias = FALSE
)
}
\arguments{
\item{signal_length}{length of the signal to be processed}

\item{dt}{sample spacing, default is 1}

\item{dj}{scale distribution parameter, default is 0.125}

\item{wavelet}{wavelet object}

\item{unbias}{whether to unbias the power spectrum, as in Liu et al. 2007.
Given a list of temporal 1D filters of variable size, this method
creates a list of `nn_conv1d()` objects that collectively form the filter bank.}
}
\description{
Continuous wavelet transform as described in Torrence and Compo,
"A Practical Guide to Wavelet Analysis".
}
\details{
Like _QUVA-Lab/PyTorchWavelets_, this does not use the FFT,
but a filter bank consisting of `torch::nn_conv1d()` modules.
The code expects input sequences to be one-dimensional. It works with batches
of signals.
}
